netlog backup unix socket setup refac 2
