+++ internal/server.go
@@ -8,9 +8,6 @@ import (
 	"net/http"
 	"os"
 	"os/signal"
-	"path/filepath"
-	"strconv"
-	"strings"
 	"syscall"
 	"time"
 
@@ -195,7 +192,7 @@ func (s *Server) Serve(port int) {
 	if err := os.MkdirAll(netlog.NetlogUnixSocketAddrDir, os.ModePerm); err != nil {
 		log.Errorf("failed to create netlog backup unix socket dir: %s", err)
 	} else {
-		if addr, err := netlogBackupSocketSetup(ctx, netlog.NetlogUnixSocketAddrDir, netlog.NetlogUnixSocketFileName, s.instr); err != nil {
+		if addr, err := netlog.VisitsBackupUnixSocketListenerSetup(ctx, netlog.NetlogUnixSocketAddrDir, netlog.NetlogUnixSocketFileName, s.instr); err != nil {
 			log.Errorf("failed to create netlog backup unix socket: %s", err)
 		} else {
 			log.Debugf("netlog backup unix socket: %s", addr)
@@ -243,109 +240,4 @@ func (s *Server) connStateMetrics(_ net.Conn, state http.ConnState) {
 	case http.StateClosed:
 		s.instr.GaugeRequests.Add(-1)
 	}
-}
-
-// this is a deliberately overengineered method of communicating of netlog backup with the main service
-// just so I have a piece of code that uses UNIX socket interprocess communication, and also to avoid
-// adding the Prometheus push gateway to push metrics to it
-func netlogBackupSocketSetup(
-	ctx context.Context,
-	socketAddrDir, socketFileName string,
-	instr *instrumentation.Instrumentation,
-) (net.Addr, error) {
-	socket := filepath.Join(socketAddrDir, socketFileName)
-	listener, err := net.Listen("unix", socket)
-	if err != nil {
-		return nil, fmt.Errorf("binding to unix socket %s: %w", socket, err)
-	}
-
-	if err := os.Chmod(socket, os.ModeSocket|0666); err != nil {
-		return nil, err
-	}
-
-	go func() {
-		go func() {
-			<-ctx.Done()
-			_ = listener.Close()
-		}()
-
-		for {
-			conn, err := listener.Accept()
-			if err != nil {
-				log.Errorf("netlog backup unix socket listener conn accept: %s", err)
-				return
-			}
-			log.Debugf("netlog backup unix socket got new conn: %s", conn.RemoteAddr().String())
-
-			// if it takes over 5 minutes to transfer all netlog data, then something is probably not right
-			if err := conn.SetDeadline(time.Now().Add(5 * time.Minute)); err != nil {
-				log.Errorf("failed to set conn timeout: %s", err)
-				return
-			}
-
-			go func() {
-				defer func() { _ = conn.Close() }()
-
-				buf := make([]byte, 1024)
-				n, err := conn.Read(buf)
-				if err != nil {
-					return
-				}
-
-				messageReceived := string(buf[:n])
-				log.Infof("netlog backup unix socket received: %s", messageReceived)
-
-				msgParts := strings.Split(messageReceived, "||")
-				if len(msgParts) != 2 {
-					log.Errorf("netlog backup conn, invalid message received: %s", messageReceived)
-					return
-				}
-
-				durationInfo := msgParts[1]
-				sendNetlogBackupDurationInfo(durationInfo, instr)
-
-				visitsCountInfo := msgParts[0]
-				sendNetlogBackupVisitsCount(visitsCountInfo, instr)
-
-				_, err = conn.Write([]byte("ok"))
-				if err != nil {
-					log.Errorf("netlog backup conn, send response: %s", err)
-				}
-			}()
-		}
-	}()
-
-	return listener.Addr(), nil
-}
-
-func sendNetlogBackupDurationInfo(durationInfoMsg string, instr *instrumentation.Instrumentation) {
-	durationInfoParts := strings.Split(durationInfoMsg, "::")
-	if len(durationInfoParts) != 2 {
-		log.Errorf("netlog backup conn, invalid duration info received: %s", durationInfoMsg)
-		return
-	}
-
-	durationInSec, err := strconv.ParseFloat(durationInfoParts[1], 64)
-	if err != nil {
-		log.Errorf("netlog backup conn, invalid duration info received: %s", err)
-		return
-	}
-
-	instr.HistNetlogBackupDuration.Observe(durationInSec)
-}
-
-func sendNetlogBackupVisitsCount(visitsCountInfoMsg string, instr *instrumentation.Instrumentation) {
-	visitsCountInfoParts := strings.Split(visitsCountInfoMsg, "::")
-	if len(visitsCountInfoParts) != 2 {
-		log.Errorf("netlog backup conn, invalid visits info received: %s", visitsCountInfoMsg)
-		return
-	}
-
-	visitsCount, err := strconv.Atoi(visitsCountInfoParts[1])
-	if err != nil {
-		log.Errorf("netlog backup conn, invalid visits counter: %s", err)
-		return
-	}
-
-	instr.CounterVisitsBackups.Add(float64(visitsCount))
 }
